move selectable cell into cell components
